# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bedrock SSO Proxy API",
    "description": "JWT-authenticated HTTP proxy server that provides secure access to AWS Bedrock APIs",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/costs": {
      "get": {
        "tags": [
          "Cost Management"
        ],
        "summary": "Get All Model Costs",
        "description": "Retrieve all model cost configurations (admin only)",
        "operationId": "get_all_model_costs",
        "responses": {
          "200": {
            "description": "List of all model costs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelCost"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Cost Management"
        ],
        "summary": "Batch Update Model Costs",
        "description": "Update multiple model costs from CSV data (admin only)",
        "operationId": "update_all_model_costs",
        "requestBody": {
          "description": "CSV content with model cost data",
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch update results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCostsResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid CSV",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
    "/api/admin/costs/{region}/{model_id}": {
      "get": {
        "tags": [
          "Cost Management"
        ],
        "summary": "Get Model Cost",
        "description": "Retrieve cost configuration for a specific model in a region (admin only)",
        "operationId": "get_model_cost",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "description": "AWS region (e.g., us-east-1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model_id",
            "in": "path",
            "description": "Model identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model cost configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCost"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model cost not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Cost Management"
        ],
        "summary": "Create or Update Model Cost",
        "description": "Create or update cost configuration for a specific model (admin only)",
        "operationId": "upsert_model_cost",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "description": "AWS region (e.g., us-east-1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model_id",
            "in": "path",
            "description": "Model identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Model cost updated successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Cost Management"
        ],
        "summary": "Delete Model Cost",
        "description": "Delete cost configuration for a specific model (admin only)",
        "operationId": "delete_model_cost",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "description": "AWS region (e.g., us-east-1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model_id",
            "in": "path",
            "description": "Model identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Model cost deleted successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "API Error response for documentation",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "ModelCost": {
        "type": "object",
        "required": [
          "id",
          "region",
          "model_id",
          "input_cost_per_1k_tokens",
          "output_cost_per_1k_tokens",
          "updated_at"
        ],
        "properties": {
          "cache_read_cost_per_1k_tokens": {
            "type": [
              "string",
              "null"
            ]
          },
          "cache_write_cost_per_1k_tokens": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "input_cost_per_1k_tokens": {
            "type": "string"
          },
          "model_id": {
            "type": "string"
          },
          "output_cost_per_1k_tokens": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ModelCostRequest": {
        "type": "object",
        "description": "Request/Response for model cost management",
        "required": [
          "region",
          "model_id",
          "input_cost_per_1k_tokens",
          "output_cost_per_1k_tokens"
        ],
        "properties": {
          "cache_read_cost_per_1k_tokens": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Optional cost per 1000 cache read tokens in USD"
          },
          "cache_write_cost_per_1k_tokens": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Optional cost per 1000 cache write tokens in USD"
          },
          "input_cost_per_1k_tokens": {
            "type": "number",
            "format": "double",
            "description": "Cost per 1000 input tokens in USD"
          },
          "model_id": {
            "type": "string",
            "description": "Model identifier (e.g., \"anthropic.claude-sonnet-4-20250514-v1:0\")"
          },
          "output_cost_per_1k_tokens": {
            "type": "number",
            "format": "double",
            "description": "Cost per 1000 output tokens in USD"
          },
          "region": {
            "type": "string",
            "description": "AWS region (e.g., \"us-east-1\")"
          }
        }
      },
...
      "UpdateCostsResult": {
        "type": "object",
        "description": "Result of updating model costs",
        "required": [
          "total_processed"
        ],
        "properties": {
          "total_processed": {
            "type": "integer",
            "description": "Number of model costs processed",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "jwt_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "x_api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}